Added send_operational_update_message() inside notification_operations domain
